Add support to adding comments to MRs from vim
This commit adds all the needed functions in order to support adding
comments into gitlab's MR directly from vim.
This is done by wrapping a `curl` command with some vimscript to make it
run from vim.

The code is not perfect and has a lot of problems that should be fixed
in the future. The major problems are presented as `TODO` comments in
the code itself.

The new code depends on some other resource. All the things that this
commit depends on presents in a comment in the README file. Currently,
this comment just says what these dependencies are. When rewriting the
README for release (issue #5) it should be updated. This information
should include information about why these functions are needed, with
explanation about how to install them and a link to their site.

Fixes #2