Refactor adding comments for more abstraction
The functions that were called whenever a new comment was added was very
specific to adding comments. Shortly, I will add the option to add
discussion threads, not just comments. To make it easier, this commit
changes the way that comments are added for more abstraction, so the
same function will be used for both times.

In order to achieve the wanted abstraction, this commit creates an enum
of the wanted command to add. The wanted command is moved between the
different functions, that can now add any kind of command. This enum
value is being parsed when the URL is being created, but aside from this
function, all the other functions works exactly the same.
Currently, the enum supports only the command of ADD_COMMENT, but it can
be extended easily.

This commit was done as part of issue #3