Add option to open discussion threads
Add the needed code in order to make it possible to open a new
discussion threads in the MR using this plugin.

This adds a new command to the plugin, that works exactly like the
command to add comments, but instead of comments it opens general
discussion threads.

This commit fixes #3