Refactor running action functions
Refactor some of the functions that run when a new action is run in
gitlab. This refactor make the code more abstract, in order to have less
duplication in it, be more extendable.

This commit does not change any functionality of the plugin.

This commit was done as part of issue #3