Change commands to get the content as json
Until now, the commands in the plugin has gotten the content of the
post action as 'body=<string>' whereas the string was the string from
the user. This worked okay for the current commands (add a comment and
add a general discussion thread), but it won't work for adding
a discussion thread on the code.
The discussion thread on the code will need to have a lot more
parameters. The easiest way to work with these parameters are to move
them as a json (which is equivalent to dicts in vimscript). In order to
make the code consistent, this commit changes the current code to use
dictionaries to move their data as well.

This commit does not change any functionality of the plugin, it just
refactors the existing code.

The commit was done as part of issue #4.