Add functionality to add code discussion thread
Add a new functionality to the plugin, the ability to add code
discussion threads into a gitlab MR. This is the last functionality that
the plugin need in order to be able to add any type of comment into
a gitlab MR.

Currently, this works like the other functionalities in the code, which
means that it asks the user (or get from him in the command line) all
the arguments the command needs. In the case of code discussion
thread, it makes the command quite awkward and hard to use (because
there are a lot of arguments). In the next versions of the plugin it
will become easier and easier to use it (example for such issues are #8,

This commit fixes #4