Add gitlab_private_token configuration variable
The variable of gitlab_private_token is added in this commit. For most
users, this variable is a single constant variable that will stay the
same forever, so it is better to have option to get it as a global.

All the old functionality stayed the same, but now, in case the value of
this global is defined, it is possible to use it instead of inserting it
for every command.

This commit was done as part of issue #14