Add cache mechanism for common values in commands
A lot of the command are running with the exact same arguments over
time.
Example for this is the different commit hashes when adding comments on
code. Whenever someone is moving over the code, all his comments will be
on the exact same hash. The cache mechanism will save the last value
that was inserted, and set it as a default value for future
commands. This will change whenever the user will insert new value
there.

This mechanism will help users add a lot of comments on the same code
without the need to insert the same values over and over again.

This commit was done as part of issue #14