Refactored similar functions to use base functions
All the exported functions that were responsible on running actions in
gitlab (add comment, add general discussion thread and add code
discussion thread) looked exactly the same:
* Check how many arguments there are
* If the number of arguments is valid:
    * Call the appropriate function
* else
    * Raise error

This commit creates a base function that is a template method that does
just that.
This removes a lot of code duplication, but it is quite awkward in
vimscript. This is awkward because I couldn't find a way to extract list
to arguments automatically, so every function that is being called that
way have an adapter function that gets the arguments as list and calls
the original function with the separated arguments.

This commit was done as part of issue #14