Add options to running actions with less arguments
Now, that we have the cache mechinsem, it is possible to add a lot
easier to use interfaces, using the arguments from the cache instead of
getting them from the user.

The most useful commands are these that can get only the text of the
body, and get all the other arguments from the cache. Currently, it is
possible only for adding comments and adding general discussion threads
(it is not possible for code discussion threads because all the
arguments of the location can't be saved in the cache).

This commit adds these command options to the plugin.

This commit was done as part of issue #14