Add option to set specific value in the cache
Add another command to the plugin MRInterfaceUpdateValueInCache. This
command get a name of a key from the cache and its new value, and update
the value of the cache for the new key.

This will let the user automate adding keys that always stay the
same. Even though the user can update the cache, it stays an internal
variable of the plugin and doesn't turn into a configurable
variable. This is done because the plugin should have full control on
this value, and the user shouldn't change it without the specific
callbacks of the plugin.

This commit was done as part of issue #14