Make the cache insert the values automatically
Change the behavior of the cache to insert the values from the cache
automatically, instead of only saving them as the default values.
It seems more convenient that way, since the user will have to type
a lot less information.

The commit also adds a flag that will enable the user to continue using
the old functionality (where the cache was just the default). By
default, this flag will make the cache be inserted automatically.

This commit was done as part of issue #14