Wrote the full README.md file
Until now, the README file had only the most basic information of the
plugin, since the plugin was not yet stable. Right now, when preparing
the plugin to release, I filled the README with all the needed
information on the plugin.

Now, the README contains any information needed in order to start using
the plugin.

This commit was done as part of issue #5