Add helpfile to the plugin
This commit adds the full helpfile of the plugin. The helpfile is the
file that will be shown whenever a user runs the help command on
commands or flags from this plugin.

The helpfile was generated using vimdoc from the code itself. It keeps
the documentation only on one place, the code, which is great. Every
time the documentation on the code changes, whoever changes it should
run the vimdoc on the code again (which will update the helpfile
accordingly).

This commit adds general documentation to the plugin, fixes some
problems in the existing documentation and run the vimdoc command to
generate the helpfile.

The commit updates the CONTRIBUTING guidelines about the vimdoc
information as well.

This commit was done as part of issue #5.