Add mutex on all the commands in the plugin
The next commits will change some commands in the plugin to get their
variables from a buffer.

The best way I found to do it, was to add a new buffer, add an
autocommand that will happen once this plugin is deleted, and wait until
the user will close the buffer.

This will mean that the user will be able to run whatever he wants in
`the middle` of commands (since the command won't really end until the
buffer will be closed). The mutex will make sure that no commands are
running in the same time. This is done in order to avoid unexpected
results.

This commit was done as part of issue #7.