Add support for getting body from buffer
This commit adds all the needed support in order to get the body of the
various commands from a buffer instead of getting it with the input
method.

This decision was made because it is easier this way to add comments
with multiple lines in them, and it easier to write long comments (all
vim's keys are working).

The way that this is done is weird, but I couldn't find a better
one. I could not find any way to open buffer and than continue from the
same location after the user has closed this buffer. This functionality
is needed in order to get the input from the user using buffers.
In order to achieve this functionality, what I did was to create
a function that runs automatically when the buffer is wiped, which will
happen automatically once the user closes the buffer.

It generates some ugly code, but overall I think that this is not so
bad, and the code is still very readable and extendable that way.

This commit fixes #7