Add documentation about inserting body into buffer
This commit adds some documentation about the behavior that will happen
when opening the new buffer for body in the various comments adding
functions.

The information is now fully in the documentation of the plugin, and it
should be easy for the user to understand how to use these buffers and
why they are there.

This commit was done as part of issue #7.