Add command to add discussion thread on current old line
Create a new command in the plugin that opens a new discussion thread on
the current line of old code.

This command makes it really easy to add comments on old lines of code
whenever you move over the code. All you need to do is run the command
when you are on the wanted line of code, and the new discussion thread
will be opened in gitlab.

Since a new command was added to the plugin, the commit also updates the
documentation of the plugin to include the needed information for the
new command.

This commit was done as part of issue #8