Reordered the code of the plugin
The code has grown a lot during the development of it. First, it was
ordered by internal and external functions only.

Now, there are lots of internal functions. This commit reorders the code
and adds a lot of different sections inside the internal functions
section. This way, it should be easier to understand where the code for
certain actions appear.

The commit also add sections to the commands file, to have each command
in its own section (since there are a lot of commands, with a lot of
documentation for every command).

This commit does not change the code at all.

This commit was done as part of issue #8