Create MRInterfaceAddCodeDiscussionThreadOnNewCode command
Create a new command in the plugin. This command will let the user add
a comment on code that has been added during the current merge request.

This command is the same as the command for old code, but it works on
new code instead. Adding it required only coping the various wrapper
methods that were for the old code adding, and updating them for the new
code.

The commit adds both the command and its implementation. As part of
adding the command, the documentation of the plugin was updated (to
include documentation on the new command as well).

This commit was done as part of issue #8