Update documentation for some commands
It appears that the commands of
MRInterfaceAddCodeDiscussionThreadOnOldCode and
MRInterfaceAddCodeDiscussionThreadOnNewCode supports also adding
comments in files that were deleted or added (respectively) during this
merge request.

It is not the behaviour I expected them to have, so I didn't add this
documentation when first writing the commands, but after checking it, it
seems that those commands work for this functionality.

This commit was done as part of issue #8