Add documentation about comments on unchanged code
It seems reasonable that there would be a command to add comments on
unchanged code. However, it is harder than it seems.

This commit adds documentation about why it is hard, and when and how it
might be supported in the future.

This commit was done as part of issue #8.