Add option to add comments on code from fugitive
Fugitive is probably the most wildely used git plugin around. It has
some commands that let the user run diff between revisions and move on
the code that way.

The name of the old files in the diff is different from their real name,
it include some kind of header to the file name that fugitive
generates.

This commit adds code that will make it possible to add comments on such
code. The commit adds code that removes all the header of fugitive for
every file name, so it will be ready to be sent to gitlab.

This option can be turned off by using the flag of
"support_fugitive_file_names".

This commit was done as part of issue #8.