Make comment with empty body not added
We don't want to add empty comments to gitlab MRs. In case the body of
the comment is empty, it means that the user has decided that he doesn't
really want to add such comment.

Until now, we had handling for this issue in case the user run a command
that got the body from the buffer. However, this did not happen for
commands that run with position automatically.
This commit fixes this issue, and make comments that have the position
in them stop in case the body is empty as well.

This commit was done as part of issue #8