Make the plugin automatically get MR id
In case the user is doing a merge request using the command of `git mr`,
it is possible to get the id of the merge request automatically from the
name of the current branch.
This commit adds functionality to the plugin to calculate automatically
the id of the merge request in such cases.

Now, whenever the plugin will need to get the merge request id, it will
do so by:
* Getting the value from the cache, in case it is there.
* Getting the value from the branch name, in case it is there.
* Getting the value directly from the user.

The plugin has flags that will allow the user to choose when he will be
asked to insert the values by himself, and when will the values be
inserted automatically.

This commit was done as part of issue #9