Add customization to order of cache and defaults
The previous commit added an option to get values from some kind of
default methods that the plugin can run by itself.
In the previous commit, the plugin always started by checking the cache,
and only in case the value was not in the cache, the default function
for the variable was called.

This commit adds a flag that will let the user customize this. The user
will be able to decide wheter he wants to strat by looking for the value
in the default values, or he wants to use the cache first.

This commit was done as part of issue #9