Add option to add all defaults to the cache
To make it easier to use the defaults in future calls, it is now
possible to set all (or some of) the defaults into the cache in a single
command.

This command will move over all the variables that can have defaults and
add the appropriate default for this variable. The command also has an
option to run for a single variable, if it is given by name.

When writing the code, I needed to have a variable that has function
references in it. To do, I had to move the variables section to the end
of the file (because of vimscript loading order).

This commit was done as part of issue #9