Refactor the external functions of the plugin
Most of the external functions had a complete code duplication between
them. All of them run the command with the given number of arguments,
surrounding it with try...catch for errors and enter...exit to not run
commands inside one another.

All this code has been moved to a new internal function. Now, most of
the external functions are just wrappers to this function that call it
with the right arguments for the given command.

This commit was done as part of issue #9