Add reference parsing to the plugin
Add a new feature to the plugin that will parse the git reference
inserted by the user and send the full one to gitlab.

This feature will allow the user to both use shorter IDs for its
commits, and use different, easier to use references to them (for
example, branch name, HEAD or tag name).

This feature can be turned on and off by a flag in the plugin.

This commit was done as part of issue #9