Add curl additional flags flag
This commit adds a new flag to the plugin's flag. This flag will let the
user custom the curl command and add more flags to it.

This can be used for various actions that are not supported by
default. For example, connect using insecure connection.

This commit was done as part of issue #37