Update README with new fixed issue
The issue of adding configured flags to curl command was fixed in the
previous commits in this branch. This commit adds a link to this issue
into the README, and update it as finished and ready as part of version
v1.0 (the next version that will be released).

This commit was done as part of issue #37