Fix escaping of quotes in body string
Until now, every time the user has inserted single quote to a string in
the body, it made the code break, and the comment was not added. This
was the case because the single (or double) quote has broken the string
of the body, and it made the CURL command to be invalid.

This commit fixes it, and make the code handle specifically the cases of
single and double quotes in the string of the body.
Handling these cases is hard - it is hard because it needs to handle the
quotes to make it compatible both with bash and with vim. Both of these
programs doesn't have easy ways to handle the strings in them, so fixing
it required a lot of research and tries. As for now, it seems to work
for the cases I have tried it with.

The current fix is still a patch, and we should find a better way to do
it in the future (or at least move it to a more generic function that
will handle all the special cases).

This works in Linux systems, but it might not work the same when the
shell is different. I don't know how it will behave under MAC-OS or
windows systems.

The command of shellescape (that should have done just that) does not
seem to work.

This commit was done as part of issue #36